refactor: rename logic.py to utils.py.
