+++ testproject/transactions/views.py
@@ -3,7 +3,7 @@ Transaction Views
 """
 from django.db.models import Q
 from rest_framework import generics, permissions, serializers
-from transactions.logic import commission_calculation, wallet_transaction
+from transactions.utils import commission_calculation, wallet_transaction
 
 from .models import Transaction
 from .serializers import TransactionSerializer
